requirements.txt
absl-py==1.1.0
appnope==0.1.3
asttokens==2.0.5
async-timeout==4.0.2
atomicwrites==1.4.1
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.3.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
cycler==0.11.0
debugpy==1.6.2
decorator==5.1.1
Deprecated==1.2.13
distlib==0.3.4
entrypoints==0.4
executing==0.9.1
filelock==3.6.0
fonttools==4.31.2
fuzzywuzzy==0.18.0
idna==3.3
iniconfig==1.1.1
ipykernel==6.15.1
ipython==8.4.0
jedi==0.18.1
jupyter-client==7.3.4
jupyter-core==4.11.1
kiwisolver==1.4.2
Levenshtein==0.20.5
lyricsgenius==3.0.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
mysql-connector-python==8.0.30
nest-asyncio==1.5.5
numpy==1.22.3
packaging==21.3
pandas==1.5.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.30
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.8.0
Pygments==2.12.0
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-Levenshtein==0.20.5
python-slugify==6.1.2
pytz==2022.4
pywin32==304
pyzmq==23.2.0
rapidfuzz==2.11.1
redis==4.3.3
requests==2.28.0
six==1.16.0
soupsieve==2.3.2.post1
spotipy==2.20.0
stack-data==0.3.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tornado==6.2
traitlets==5.3.0
typing_extensions==4.2.0
urllib3==1.26.9
virtualenv==20.14.0
wcwidth==0.2.5
wrapt==1.14.1